Fitness_For_EveryOne/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyJWT>=2.8.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
